feat(PrintService): Implement multiframe printing support and refactor related methods refactor(BatchPrintClient): Update job data structure for multiframe printing refactor(services): Remove BatchPrintService and update service references
