+++ src/Mapbender/PrintBundle/Resources/config/elements.xml
@@ -28,7 +28,7 @@
             <argument type="service" id="security.token_storage" />
             <argument type="service" id="mapbender.source.url_processor.service" />
             <argument type="service" id="mapbender.print.template_parser.service" />
-            <argument type="service" id="mapbender.batch_print.service" />
+            <argument type="service" id="mapbender.print.service" />
             <argument type="service" id="mapbender.print_plugin_host.service" />
             <argument>%mapbender.print.memory_limit%</argument>
             <argument>%mapbender.print.queueable%</argument>
